+++ packages/core/src/index.ts
@@ -64,6 +64,7 @@ export * from "./resolver/Semver";
 export * from "./resolver/Types";
 export * from "./support/Execute";
 export * from "./support/Functional";
+export * from "./support/Integrity";
 export * from "./support/Json";
 export * from "./support/Log";
 export * from "./support/Platform";